Arreglos para visualizar bien el porcentaje de uso de la grid

+++ src/main/java/com/example/restproyect/mocks/Mockgrid.java
@@ -53,7 +53,7 @@ public class Mockgrid {
 	}
 
 	public double getWorkload() {
-		return 1 - (nodosDisponibles / CANTIDAD_NODOS);
+		return 1 - ((double)nodosDisponibles / (double)CANTIDAD_NODOS);
 	}
 	
 	public void procesarSimulacion(Documento documento, ColaPaquete paquetes) {
